Refactored search subcommand into separate modules
