+++ packages/cli/lib/config.js
@@ -1,0 +1,4 @@
+// lib/config.js
+
+export const excludedDirs = ["dist", "node_modules"];
+export const searchRegexFlags = "gi";
